refactor: common 폴더 index.ts 파일 추가
