+++ src/components/Map/index.tsx
@@ -17,8 +17,7 @@ import { useQuery } from '@tanstack/react-query';
 import { useEffect, useRef, useState, useCallback } from 'react';
 import { getSidoDustInfos, getCityDustInfos } from '@/apis/dustInfo';
 import { getAllLocation } from '@/apis/location';
-import DustLevel from '@/components/common/DustLevel';
-import DustState from '@/components/common/DustState';
+import { DustLevel, DustState } from '@/components/common';
 import useMap from '@/hooks/useMap';
 import theme from '@/styles/theme';
 import type { CityDustInfo, MarkerInfo } from '@/types/dust';
