+++ src/pages/DustForecast.tsx
@@ -3,8 +3,7 @@ import { useQuery } from '@tanstack/react-query';
 import { motion } from 'framer-motion';
 import { useSearchParams } from 'react-router-dom';
 import { getCityDustInfos } from '@/apis/dustInfo';
-import AsyncBoundary from '@/components/common/AsyncBoundary';
-import DustLevel from '@/components/common/DustLevel';
+import { AsyncBoundary, DustLevel } from '@/components/common';
 import CurrentDustInfo from '@/components/DustForcast/CurrentDustInfo';
 import DustChart from '@/components/DustForcast/DustChart';
 import ForcastInfo from '@/components/DustForcast/ForcastInfo';
