+++ src/pages/Ranking.tsx
@@ -4,9 +4,7 @@ import { motion } from 'framer-motion';
 import { ChangeEvent, useState, useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 import { getSidoDustInfo } from '@/apis/dustInfo';
-import AsyncBoundary from '@/components/common/AsyncBoundary';
-import DustFigureBar from '@/components/common/DustFigureBar';
-import DustState from '@/components/common/DustState';
+import { AsyncBoundary, DustFigureBar, DustState } from '@/components/common';
 import SelectList from '@/components/Ranking/SelectList';
 import SidoRankList from '@/components/Ranking/SidoRankList';
 import theme from '@/styles/theme';
